Isolate unit test covering custom methods of storage entity transaction
